# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Otacon Phone Automation API",
        "description": "REST API for controlling Android phones connected to a Raspberry Pi.",
        "license": {
            "name": "MIT"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "https://otacon-pi.<tailnet>.ts.net:8080",
            "description": "Default Tailscale address"
        }
    ],
    "paths": {
...
        "/api/wifi": {
            "get": {
                "tags": [
                    "WiFi"
                ],
                "operationId": "getWifiStatus",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WifiStatus"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "WiFi"
                ],
                "operationId": "setWifiEnabled",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SetWifiEnabledBody"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WifiStatus"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "SetWifiEnabledBody": {
                "type": "object",
                "required": [
                    "enabled"
                ],
                "properties": {
                    "enabled": {
                        "type": "boolean"
                    }
                }
            },
...
            "WifiStatus": {
                "type": "object",
                "required": [
                    "enabled",
                    "connected",
                    "desired_enabled"
                ],
                "properties": {
                    "connected": {
                        "type": "boolean"
                    },
                    "desired_enabled": {
                        "type": "boolean",
                        "description": "Persisted host-local desired state for this phone."
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "rssi": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int32"
                    },
                    "ssid": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                }
            }
        }
    }
}